requirements.tests.txt
django-crypto-fields
django_celery_beat
django_celery_results
edc_adverse_event
edc_appointment
edc_auth
edc_dashboard
edc_data_manager
edc_device
edc_export
edc_identifier
edc_lab
edc_lab_dashboard
edc_list_data
edc_locator
edc_metadata
edc_metadata_rules
edc_navbar
edc_notification
edc_offstudy
edc_pharmacy
edc_randomization
edc_reference
edc_registration
edc_sites
edc_test_utils
edc_visit_schedule
faker
inte_ae
inte_auth
inte_dashboard
inte_consent
inte_export
inte_form_validators
inte_labs
inte_lists
inte_metadata_rules
inte_prn
inte_reference
inte_screening
inte_sites
inte_subject
inte_visit_schedule
model-bakery
